requirements.txt
streamlit
google-generativeai
python-dotenv
graphviz
Pillow
google-cloud-vision
openai